requirements.txt
Flask==3.0.2
twilio==9.0.2
anthropic>=0.49.0
httpx==0.27.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==21.2.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
APScheduler==3.10.4
pytz
